backend/services/auth-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic[email]==2.5.3
email-validator==2.3.0
dnspython==2.8.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
python-jose[cryptography]
bcrypt>=4.0.0
pyjwt>=2.8.0
google-auth
google-auth-oauthlib
requests==2.31.0
cryptography>=41.0.7

